requirements.txt
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.59.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1